Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > > Patch Set 1:
> > > > > 
> > > > > How about we keep everything as is, but just do an os.MkdirAll('testdata', 0644) before reading/writing the file ?
> > > > > 
> > > > > That should make it work in the playground and make a very minimal change.
> > > > 
> > > > 
> > > > It would work for WriteFile, but not for ReadFile, you still have to creat the file.
> > > 
> > > Indeed. I don't have a better idea for ReadFile. Even os.Open() does not have an example.
> > 
> > The WriteFile example should not be writing to the 'testdata' directory anyway; see #28387.
> > 
> > I also don't understand why `golang.org/pkg/io/ioutil' is presenting the examples as a `func main()` in the first place. There is no general requirement that `Example` functions be runnable as standalone `main` functions; if we want to convert Example functions to run in the Playground we should arguably convert them as the entire _test.go file, perhaps with the other tests and examples in the file redacted out and a fresh `goimports` run to clear out any additional dependencies that are no longer needed.
> 
> I also see https://golang.org/pkg/os presents example in `main()`, or I missed something?

To answer my own question: the use of `package main` for examples was introduced in CL 6525046, and the rationale (and intended behavior) for it is not clearly documented in that change.

However, it does appear that the example-extraction logic there includes any top-level declarations required by the example body. So one option might be, if not to separate the boilerplate, to at least reduce it.

For example, we could have a top level `var tmpDir` string, left blank for the Playground but initialized to a temporary directory before a call to testing.TestMain otherwise.

Patch-set: 1
Reviewer: Gerrit User 6365 <6365@62eb7196-b449-3ce5-99f1-c037f21e1705>