net/http/httptrace: create internal copy of httptrace.ClientTrace in httptrace.WithClientTrace
In case of multiple calls to httptrace.WithClientTrace with different, or even
the same, input httptrace.ClientTrace object. The API is designed in such
a way that hook functions that are defined by older trace objects were all invoked
in reverse order (from newer to older one). This was implemented by modifying
the fields of newest httptrace.ClientTrace object with new functions that invoked
both older and newest hook function. This causes stack overflow and inconsitency
of number of times hook functions are called if the same httptrace.ClientTrace
object is registered multiple times.

Fixes #32925

Change-Id: Iefa8b7fa4ac88f7d518ef9c035e57169e6a1268c