Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> A change like this needs a test case, likely somewhere under misc/cgo.

I've updated the CL with a new test for `go test cmd/go`: it doesn't really rely on cgo, other than forcing external link (and the test can be written without it, using the build option -ldflags='-linkmode=external' if needed).  The test passes with the changes to cmd/link, and fails when those same changes are backed out.

Patch-set: 2
CC: Gerrit User 32979 <32979@62eb7196-b449-3ce5-99f1-c037f21e1705>