cmd/go: fix "go version" bug #31861
On ELF, the "go version" command assumes that the .go.buildinfo
section is the first read-write data segment in the binary.  This is
not the case when an external linker was used.

Look up the section by name instead.

Fixes #31861

Change-Id: I5b3c34d5c49329bcff9414adc8be315af77482b8
GitHub-Last-Rev: 7075969868d7a7e460f04589de9ac01897ae1be4
GitHub-Pull-Request: golang/go#33172