Update patch set 1
Patch Set 1:

Okay.  I tried to add a test, but I need some help with that.

I added a second check to testdata/script/version.txt, but it doesn't look like this test case builds with cgo.  When I build Go from git locally, it does have cgo support.

I ran the test case like this:

~/git/go/src$ ../bin/go test -v cmd/go -run=Script/^version$

While running the newly added part, the "go build" command emits a warning to stderr:

            > go build -ldflags=-linkmode=external -o external.exe rsc.io/fortune
            [stderr]
            # rsc.io/fortune
            loadinternal: cannot find runtime/cgo
            >

I think it fell back to internal linking, so it won't be an effective demonstration of the fix.  Is that normal?  Is there a better way to run the test case?

Patch-set: 1
CC: Gerrit User 33002 <33002@62eb7196-b449-3ce5-99f1-c037f21e1705>