path: This change enforces the pattern syntax evaluation before the matching process on Match function FIX #28614
The current implementation of the glob in `path.Match(pattern, name)` function is evaluating the pattern syntax in a lazy way, this can lead to some miss confusions or mistakes, due to the same invalid pattern, can end up without match or with a syntax error depending on the input.

Note: If this solution is good enough, I'll apply it to the `filepath.Match(pattern, name)` function too.

Fixes #28614

Change-Id: I055a3b707b93f8ddda87551effb90f5e767a4117
GitHub-Last-Rev: 209905200982d8dbc0b27e039541950750f63e12
GitHub-Pull-Request: golang/go#33189