cmd/compile: Move debug_line generation from linker into compiler.
This is the first step in what will be a stack of CLs to move debug_line
generation from the linker into the compiler. This change first:

 - Creates a new symbol for debug_lines on each function object
   - Eventually, the is_stmt symbol will be eliminated, preserving
     conservation of symbols.
 - Populates that symbol with the debug_lines information.

Future changes will actually use this new data in the linker, eliminate
is_stmt, and do some additional cleanup in the linker (there's likely
a small amount of bookkeeping code that can probably be cleaned up
from the linker).

Change-Id: I5234720960720259e24953a2b9342353a460060c