runtime: bound small object sweeping to 100 spans
DO NOT REVIEW. This is an experiment.

Currently, the small object sweeper will sweep until it finds a free
slot or there are no more spans of that size class to sweep. In dense
heaps, this can cause sweeping for a given size class to take
unbounded time, and gets worse with larger heaps.

This CL limits the small object sweeper to try at most 100 spans
before giving up and allocating a fresh span. Since it's already shown
that 100 spans are completely full at that point, the space overhead
of this fresh span is at most 1%.

Updates #18155.

Change-Id: I381618f38c252ac421707becb4a2ecd0f07f88b2