Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)

I confess that I was ignorant of the difference between upper case and title case mapping before reading the link as it seems confined to languages other than English. Based on an informal poll of some of my colleagues it seems like this might be common.

If we accept that it's a helpful clarification, the options I see are:
1. Including a link to help add context.
2. Embed the relevant text from the link.
3. Add a shorter version of the relevant text explaining.

That said, i'm open to feedback or suggestions.

Patch-set: 2
Reviewer: Gerrit User 33094 <33094@62eb7196-b449-3ce5-99f1-c037f21e1705>