Update patch set 1
Patch Set 1:

Note that this code was intentionally written to not use rawstringtmp in https://go-review.googlesource.com/37791 for a ~20% performance improvement as string conversions are frequently used Go construct. We should likely add a comment in the current code to avoid any accidental de-optimizations.

Please provide/create a github issue (https://github.com/golang/go/issues/new) with an example program showing and explaining the mentioned race issue and if not present please include a test in this CL that would have failed before and passes now.

Patch-set: 1
CC: Gerrit User 16006 <16006@62eb7196-b449-3ce5-99f1-c037f21e1705>