Update patch set 1
Patch Set 1:

I do not see how there can be a race on buf here since buf will always be allocated by the Go compiler on the go routines stack (not on the heap) and it is unclear how another go routine would be able to read or write to that backing buffer concurrently. If that is possible other functions such as slicerunetostring would have the same problem therefor it would be helpful to provide a test program that showcases this race.

slicebytetostringtmp seems to needs to check b for races because slicebytetostringtmp is also used by some string conversions (e.g. map[string(k)]) with direct calls emitted to it by the compiler (in race mode) and there b can be allocated on the heap and concurrently used.

Patch-set: 1