Update patch set 1
Patch Set 1:

actually, I didn't know when can the concurrently read or write happen in function slicebytetostringtmp.  but I thought it was the same with slicebytetostring.  Now I get it.

thanks for your answer. I agree with you that we can avoid add race check if there won't be any race.

Patch-set: 1
CC: Gerrit User 33132 <33132@62eb7196-b449-3ce5-99f1-c037f21e1705>