Update patch set 1
Patch Set 1:

(4 comments)

Awesome, thank you David and welcome to the Go project!

Good spot for tests that need some refactoring and hardening.
Given that are making this test better, I have some suggestions that'll
harden it but also ensure that we are actually capturing desired errors such as instead making the test:


tests := []struct{
   name string
   wantErr string
   header Header
}{
   {....}
}

for _, tt := range tests {
    t.Run(tt.name, func(t *testing.T) {
        req := &Request{
              Method: "POST",
              Header: tt.header,
              Body: ioutil.NopCloser(strings.NewReader("body")),
        }

        err := req.ParseForm()
        if tt.wantErr != "" {
           if err == nil || !strings.Contains(err.Error(), tt.wantErr) {
              t.Fatalf("Error mismatch:\nGot:  %v\nWant: %v", err, tt.wantErr)
           }
        } else if err != nil {
           t.Fatalf("Unexpected error: %v", err)
        }
    })
}


Also we'll have to wait until Go1.14 for this change to go in, but nonetheless we are very close to Go1.14 and this change shall be submitted early in the cycle.

Patch-set: 1