cmd/compile: remove statement marks from secondary calls
Calls are code-generated in an alternate path that inherits
its positions from values, not from *SSAGenState.  The
default position on *SSAGenState was marke as not-a-statement,
but this was not applied to the value itself, leading to
spurious "is statement" marks in the output (convention:
after code generation in the compiler, everything is either
definitely a statement or definitely not a statement, nothing
is in the undetermined state).

This CL causes a 157 statement regression in ssa/stmtlines_test,
which is within bounds but needs to be looked at.

Change-Id: Ie9a507bd5447e906b35bbd098e3295211df2ae01