cmd/compile: pick position of implicit break statements more carefully
The previous version used the position of the switch statement,
which makes for potentially jumpy stepping and introduces a large
number of statements repeating the line (tricky for inserting
breaks).  It also shared a single OBREAK node and this was not
really a syntax "tree".

This reduces the number of failing lines in ssa/stmtlines_test by 7

Change-Id: I1f71b231a26f152bdb6ce9bc8f95828bb222f665