cmd/compile: generate debug_lines in compiler
This is mostly a copy-paste jobs from the linker to generate the
debug information in the compiler instead of the linker. The new
data is inserted into the debug line numbers symbol in the CL up
the stack.

Right now, if we were to diff the output of this CL and the
debug_lines section in the linker, there will be differences
in the DW_LNS_set_file opcodes. This is the because the linker
eliminates filesnames from the DWARF that have been dead-code
eliminated. I will fix this difference between the two in a
separate CL in the stack.

Change-Id: If9103eda6cc5f1f7a11e7e1a97184a060a4ad7fb