cmd/compile: generate debug_lines in compiler
This is mostly a copy-paste jobs from the linker to generate the debug
information in the compiler instead of the linker. The new data is
inserted into the debug line numbers symbol in the CL up the stack.

Right now, if we were to diff the output of this CL and the debug_lines
section in the linker, there will be differences.  The problem comes as
a result of the fact that the linker uses a full CU when generating the
opcode table. This change only generates for a single function at a
time. The easiest way to reconcile the two segments is to reset the
state machine in the debug sections generated from the linker. Another
CL in this stack will address that (by resetting the state machine after
every function), and comparing the output between the two approaches.

Change-Id: If9103eda6cc5f1f7a11e7e1a97184a060a4ad7fb