cmd/compile: generate debug_lines in compiler
This is mostly a copy-paste jobs from the linker to generate the
debug information in the compiler instead of the linker. The new
data is inserted into the debug line numbers symbol in the CL up
the stack.

These line numbers will have incorrect file numbers -- something
the linker will have to clean up. Specifically, the linker works on
package-level CUs, and the compiler doesn't have the complete
package in this change's location. As such, the linker will still
fixup the line number table when combining all the functions into
the CU. That change will be later on the stack.

Change-Id: If9103eda6cc5f1f7a11e7e1a97184a060a4ad7fb