cmd/compile: generate debug_lines in compiler
This is mostly a copy-paste jobs from the linker to generate the debug
information in the compiler instead of the linker. The new data is
inserted into the debug line numbers symbol defined in CL 188238.

Generating the debug information BEFORE deadcode results in one subtle
difference, and that is that the state machine needs to be reset at the
end of every function's debug line table. The reasoning is that
generating the table AFTER dead code allows the producer and consumer of
the table to agree on the state of the state machine, and since these
blocks will (eventually) be concatenated in the linker, we don't KNOW
the state of the state machine unless we reset it. So,
generateDebugLinesSymbol resets the state machine at the end of every
function.

Right now, we don't do anything with this line information, or the file
table -- we just populate the symbols.

TODO:
****

- Rewrite the filenames in the file tables from createDebugLinesTable.
  (The filenames written in that function will include bogus paths.)

- Generate the debug information from these two new symbols.

- And eventually (possibly) eliminate the is_stmt symbol (from
FuncInfo), as I think it's only used to generate the debug lines state
machine. (See further discussion in CL 188238).

- Benchmark this set of changes.
  Profiling says that removing this generation from the linker will
result in improvements to the link speed

- Investigate further cleanup.
  I think there might be other things from FuncInfo that can be
eliminated (specficially things like the file name, which I think is
only used [again] to generate the debug lines information). Again,
investigation is required here, but I suspect there's more cleanup to be
achieved.

Change-Id: If9103eda6cc5f1f7a11e7e1a97184a060a4ad7fb