cmd/gofmt: fix computation of function header size
Function sizes are computed to determine whether a function
can be kept on one line or should be split to several lines. Part of the
computation is the function header from the FUNC token and until the
opening { token.

Prior to this change, the function header size took into account the
original source position of the current token, which led to issues like
position, so that header size represents the reformatted source rather
than the original source.

Fixes #28082

Change-Id: Ib737f6933e09b79e83715211421d5262b366ec93