Update patch set 4
Patch Set 4: Run-TryBot+1

This should be ready for review, but reviewers might want to hold off.

A few other changes in the stack will modify the linker such that we can diff the output of this diff, and ensure the two outputs agree. In other words, test that this change is kosher. If parts of this diff are wrong, they'll need to be modified, etc...

I leave it up to the reviewer to make the final decision. :)

Patch-set: 4
Reviewer: Gerrit User 32560 <32560@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Run-TryBot=+1
Work-in-progress: false