sync/atomic define sync/atomic memory models
Current Go documentation does not specify any memory order guarantees
made by the atomic operations, but implementations of sync/atomic are
concerned with the ordering. So it is better to state the behavior of
the atomic operations wrt the memory model to avoid wrong assumptions
being made by Go developers.

After discussing with golang-dev mailing list to decide on what docs we
want, this CL gives the final specification.

Please refer to the detailed discussion:
https://groups.google.com/d/msg/golang-dev/vVkH_9fl1D8/azJa10lkAwAJ

Update #5045

Change-Id: I55513aaa8b41230a292749e131cfaa5272b16992