Update patch set 1
Patch Set 1:

Thanks for the change Pantelis. And congrats on your first contribution :)

If you didn't know, we have a pending issue to track the addition of examples to the regexp package. Please add "Updates #21450" to this issue and the other one. And if you want, you can go ahead and add the other examples listed here https://github.com/golang/go/issues/21450#issuecomment-440132820 to close out the issue.

Patch-set: 1
CC: Gerrit User 24096 <24096@62eb7196-b449-3ce5-99f1-c037f21e1705>