cmd/link: add notion of multiple compliation units per package
As we move the debug_line generation into the compiler, we need to
upgrade the notion of compilationUnit to not just be on a per package
basis. The reasoning for this is that the with the current notion of
compilationUnits is that they represent an entire package. That won't be
the case as it will be impossible for all compilationUnits to have the
same set of files names used to build the debug_lines table. (For
example, assembled files in a package don't know about any files but
themselves, so the debug_lines table could only reference themseves. As
such, we need to break the 1:1 relationship between compUnit and
package.)

This is a non-functional change.

Change-Id: I2e517bb6c01de0115bbf777af828a2fe59c09ce8