cmd/compile: fix typefmt has side effects on interface type
The root cause of #33658 is that fmt.Printf does have side effects when
printing Type

typefmt for TINTER will call Type.Fields to get all embedded fields and
methods. The thing is that type.Fields itself will call Dowidth, which will
expand the embedded interface, make it non-embedded anymore.

To fix it, we add a method to get list of interface fields without
expanding embedded interface, and use it in typefmt.

Fixes #33658

Change-Id: Id4b70ff68a3b802675deae96793fdb8f7ef1a4a7