Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> Are you positive about this? I seem to remember that json.Unmarshaler isn't supported for map keys, as map keys can only be strings, and not other JSON values like objects or numbers.


Map keys can be of any type that is comparable.
For example in this snippet -- https://play.golang.org/p/b7ZUS4lAmVb -- you see a struct being used as a key, and with proper methods defined UnmarshalJSON is called.

What the snippet also demonstrates is the other point made in the original issue linked to this CL. When both UnmarshalText and UnmarshalJSON are defined, the latter is called. But when only UnmarshalJSON is defined, it's not called.

This seems to be a bug because https://github.com/golang/go/blob/dcb84828a6b1d7d4d9186bc6d752cc147e255162/src/encoding/json/decode.go#L716 checks for TextUnmarshaler to check the type of the map key, but then later in https://github.com/golang/go/blob/dcb84828a6b1d7d4d9186bc6d752cc147e255162/src/encoding/json/decode.go#L933 indirect() will return a non-nil value if json.Unmarshaler is implemented.

We might add to the doc saying that TextUnmarshaler has to be defined for the custom unmarshaling to work, but then UnmarshalJSON - if it exists - will be prioritized over it. This is quirky but at least it describes what the implementation is doing, which I'm not sure we can change at this point because it will break users.

WDYT?

Patch-set: 1
Reviewer: Gerrit User 30696 <30696@62eb7196-b449-3ce5-99f1-c037f21e1705>