cmd/internal/obj: debug mode to call function on stack check
This adds support for calling a function unconditionally whenever
there is a conditional call to runtime.morestack. We'll use this to
record when constructing the runtime lock graph to record that the
function could have acquired the locks involved in stack growth under
the right circumstances, even if this particular call did not.

Change-Id: I987c50d313c4885bc65490dfb25a44c3aab289c5