cmd/compile: update object file format for DWARF file table
In CL 188317, we generate the debug_lines in the compiler, and created a
new symbol to hold the file table. Using a symbol was a mistake and
added lots of complexity. Rather, much easier was just changing the
objectfile format. Here I output the file table to a new object file
format, and read it back in the linker. The next change will move the
.debug_lines over to the new structure -- and will show any speedups
this approach has.

Change-Id: Ibee192e80b86ff6af36467a0b1c26ee747dfee37