runtime: warn instead of failing on stale runtime
Currently, the runtime tests that build an external binary fail if the
go command says the installed runtime.a is stale. This check was put
in long ago when the go command avoided rebuilding std packages, and
hence would build these external binaries with the stale runtime.a.

We no longer need this logic: now the go command will rebuild (and
cache) anything that needs to be rebuilt, including the runtime, so
just calling "go build" for these external test binaries will create
an up-to-date binary. However, we keep the logic and demote it to a
warning, since this still causes the test to take longer than it
otherwise would ("go test -short" takes 16% longer on my laptop;
increasing from 24.5s to 28.5s) and may be unexpected.

Beyond allowing the test to be run without "go install"ing the
runtime, this makes it easy to change tags and build flags when
running the runtime test.

Change-Id: I4cb27c6f03143044a7614c060e2b6fdffd2e1640